requirements.txt
Pillow==9.5.0
octoai-sdk==0.5.1
streamlit==1.27.2